+++ iot-e2e-tests/common/src/test/java/tests/integration/com/microsoft/azure/sdk/iot/digitaltwin/helpers/E2ETestConstants.java
@@ -1,6 +1,6 @@
 package tests.integration.com.microsoft.azure.sdk.iot.digitaltwin.helpers;
 
 public class E2ETestConstants {
-    public static final String IOT_HUB_CONNECTION_STRING_ENV_VAR_NAME = "IOTHUB_CONNECTION_STRING";
+    public static final String IOTHUB_CONNECTION_STRING_ENV_VAR_NAME = "IOTHUB_CONNECTION_STRING";
     public static final String MODEL_ID = "dtmi:com:example:Thermostat;1";
 }
